requirements.txt
python-telegram-bot
google-generativeai
PyPDF2
python-docx
python-dotenv 